pyproject.toml
[project]
name = "custommcpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "pydantic-ai>=1.70.0",
    "python-dotenv>=1.2.2",
]
